hide Result implementations and remove their toString overrides
